requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
celery>=5.3.0
redis>=4.5.0
requests>=2.31.0
PyJWT>=2.8.0
cryptography>=41.0.0
pywebpush>=1.14.0
cffi>=1.15.0
psycopg2-binary>=2.9.0 # Required for PostgreSQL. Use 'psycopg2' if compiling from source.
python-dotenv>=1.0.0 # For managing environment variables
gunicorn>=21.0.0 # For production deployment
django-celery-results>=2.5.0 # For storing Celery task results in Django DB
